\name{Barley}
\alias{Barley}
\docType{data}
\title{
Barley data (1973)
}
\description{
Sample data on ear per plant, ear length (cms), 100-grain weight (g), and grain yield per plant (kg) from 1973 Barley performance trial, comprising of eight varieties in four replication (Chaudhary, 1973).
}
\usage{data("Barley")}
\format{
  A data frame with 32 observations on the following 4 variables - EPR, EL,GW,and GY.
}
\source{
Chaudhary BD. (1973). Estimation of genetic parameters in barley (Hordeum vulgare L.) by diallel analysis and its modifications. Unpublished Ph.D. Thesis, Haryana Agricultural University, Hisar.
}
\examples{
data(Barley)
}
